+++ packaging/wcm/webapp/src/main/resources/locale/ecm/dialogs_en.xml
@@ -20,7 +20,7 @@
       </label>
     </dialog>
   </Ecm>
-
+  
   <UIDocumentForm>
     <label>
       <title>#{word.title}</title>
@@ -29,10 +29,10 @@
       <content>#{word.content}</content>
       <contentPlain>#{word.content}</contentPlain>
       <name>Name</name>
-      <categories>Categories</categories>
+    <categories>Categories</categories>
     </label>
   </UIDocumentForm>
-
+ 
   <UICommentForm>
     <label>
       <placeholder>Use @ to mention someone in your comment</placeholder>
@@ -287,7 +287,6 @@
   </NumberFormatValidator>
   <!--
   #####Since version 2.4#######
-
   ###################### resoucebundle exo:taxonomyAction dialogs template ####
   -->
   <ScriptAction>
@@ -303,33 +302,32 @@
     </dialog>
   </ScriptAction>
   <UIDialogForm>
-    <form>
-      <DialogFormException>Cannot save this document. There may be a problem with the pre-save interceptor.</DialogFormException>
-    </form>
+   <form>
+   <DialogFormException>Cannot save this document. There may be a problem with the pre-save interceptor.</DialogFormException>
+   </form>
   </UIDialogForm>
 
   <!--
   #####Since version 2.6#######
-
   ###################### resoucebundle exo:taxonomyAction dialogs template ####
   -->
-  <no>No</no>
-  <yes>Yes</yes>
-  <true>True</true>
-  <false>False</false>
-  <Electronic>Electronic</Electronic>
-  <Paper>Paper</Paper>
-  <NONE>None</NONE>
-  <CLASSIFIED>Classified</CLASSIFIED>
-  <UNCLASSIFIED>Unclassified</UNCLASSIFIED>
-  <NOCONTRACT>No Contract</NOCONTRACT>
-  <one-minute>One Minute</one-minute>
-  <hourly>Hourly</hourly>
-  <daily>Daily</daily>
-  <quarterly>Quarterly</quarterly>
-  <yearly>Yearly</yearly>
-  <ten-years>Ten Years</ten-years>
-  <read>User Action</read>
+<no>No</no>
+<yes>Yes</yes>
+<true>True</true>
+<false>False</false>
+<Electronic>Electronic</Electronic>
+<Paper>Paper</Paper>
+<NONE>None</NONE>
+<CLASSIFIED>Classified</CLASSIFIED>
+<UNCLASSIFIED>Unclassified</UNCLASSIFIED>
+<NOCONTRACT>No Contract</NOCONTRACT>
+<one-minute>One Minute</one-minute>
+<hourly>Hourly</hourly>
+<daily>Daily</daily>
+<quarterly>Quarterly</quarterly>
+<yearly>Yearly</yearly>
+<ten-years>Ten Years</ten-years>
+<read>User Action</read>
   <schedule>Scheduler</schedule>
   <node_added>Content Addition</node_added>
   <node_removed>Content Removal</node_removed>
@@ -366,7 +364,7 @@
       </label>
     </dialog>
   </EnableVersioning>
-
+  
   <AutoVersioning>
     <dialog>
       <label>
@@ -378,20 +376,19 @@
   </AutoVersioning>
 
   <AccessibleMedia>
-    <dialog>
-      <label>
-        <captions>Captions</captions>
-        <audioDescription>Audio description</audioDescription>
-        <alternativeText>Alternative text</alternativeText>
-      </label>
-    </dialog>
-    <help>
-      <CaptionsHelp>Captions should provide text of dialog and important sounds</CaptionsHelp>
-      <AudioDescriptionHelp>An audio description is a narrated track of the video</AudioDescriptionHelp>
-      <AlternativeTextHelp>Provide an accessible, alternative way of presenting the information</AlternativeTextHelp>
-    </help>
-  </AccessibleMedia>
-
+  <dialog>
+  <label>
+  <captions>Captions</captions>
+<audioDescription>Audio description</audioDescription>
+<alternativeText>Alternative text</alternativeText>
+</label>
+</dialog>
+<help>
+<CaptionsHelp>Captions should provide text of dialog and important sounds</CaptionsHelp>
+<AudioDescriptionHelp>An audio description is a narrated track of the video</AudioDescriptionHelp>
+<AlternativeTextHelp>Provide an accessible, alternative way of presenting the information</AlternativeTextHelp>
+</help>
+</AccessibleMedia>
 
  <!-- New Document i18n -->
   <UIActionBar>
@@ -428,5 +425,5 @@
       </not-allow-sameNameSibling>
     </msg>
   </UINewDocumentForm>
-
+  
 </bundle>
